pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true  # Important: Enables inclusion of non-code files
packages = ["pydbzengine"]

[project]
name = "pydbzengine"
version = "3.0.7.1"
authors = [
    { name = "Memiiso Organization" },
]
description = "Python Debezium Embedded Engine"
readme = "README.md"
requires-python = ">=3.8"
keywords = ["Debezium", "Replication", "CDC"]
license = { text = "Apache License 2.0" }
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "pyjnius==1.6.1",
    "dlt>=1.5.0"
]
[project.optional-dependencies]
dev = [
    "testcontainers>=4.9.1",
    "dlt[duckdb]>=1.5.0",
    "apache-airflow>=2.10.4"
]
#[project.scripts]
#debezium = "pydbzengine.__main__:main"

[project.urls]
Homepage = "https://github.com/memiiso/pydbzengine"
Documentation = "https://github.com/memiiso/pydbzengine"
Repository = "https://github.com/memiiso/pydbzengine"
